% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_counts.R
\name{filter_low_counts}
\alias{filter_low_counts}
\title{Filters out genes with low counts based on CPM threshold}
\usage{
filter_low_counts(data, one_cpm_to_counts = 10, first_col = 1)
}
\arguments{
\item{data}{counts-based dataframe of gene expression (genes x samples),
gene names in rownames}

\item{one_cpm_to_counts}{scale 1 cpm to default 10 counts}

\item{first_col}{number of first column with counts if other columns precede
columns of samples}
}
\value{
returns filtered dataframe
}
\description{
Filters out genes with low counts based on CPM threshold
}
\examples{
filtered_data <- filter_low_counts(data)
}
