% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%nin\%}
\alias{\%nin\%}
\title{not in}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{vector or NULL: the values to be matched}

\item{y}{vector or NULL: the values to be matched against}
}
\value{
if elements of \code{x} are in \code{y}
}
\description{
return true if x not in y
}
\examples{
letters[1:10] \%!in\% letters[1:3] # 30\% of the second arg ar in the first
}
